requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptg==0.2.post4
decorator==4.4.2
future==0.18.2
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
hachoir==3.1.2
httplib2==0.19.1
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
iniconfig==1.1.1
instabot==0.117.0
mock==4.0.3
MouseInfo==0.1.3
moviepy==1.0.3
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
proglog==0.1.9
protobuf==3.15.8
py==1.10.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.20
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
PySocks==1.7.1
pytest==6.2.2
pytweening==1.0.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.12.1
rsa==4.7.2
schedule==1.0.0
scipy==1.6.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
telegram-upload==0.4.0
Telethon==1.21.1
toml==0.10.2
tqdm==4.59.0
uritemplate==3.0.1
urllib3==1.26.3
